엔티티, 도메인, dto 다 있게 만들기로 확정, 문제는 dto 변환
