+++ src/main/java/com/sprint/mission/discodeit/domain/ChannelType.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.domain;
 
 public enum ChannelType {
     PUBLIC, PRIVATE
