+++ src/main/java/com/sprint/mission/discodeit/domain/User.java
@@ -1,5 +1,7 @@
 package com.sprint.mission.discodeit.domain;
 
+import com.sprint.mission.discodeit.domain.exception.ErrorType;
+import com.sprint.mission.discodeit.domain.exception.ValidationException;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -64,6 +66,22 @@ public class User {
         this.lastActiveAt=lastActiveAt;
     }
 
+    private void validateEmail(String email) {
+        if (!email.matches("^[A-Za-z0-9+_.-]+@(.+)$")) {
+            throw new ValidationException(ErrorType.INVALID_EMAIL);
+        }
+    }
 
+    private void validatePassword(String password) {
+        if (password == null || password.length() < 4) {
+            throw new ValidationException(ErrorType.INVALID_PASSWORD);
+        }
+    }
+
+    private void validateUsername(String username) {
+        if (username == null || username.isBlank()) {
+            throw new ValidationException(ErrorType.INVALID_USERNAME);
+        }
+    }
 
 }
