+++ src/main/java/com/sprint/mission/discodeit/entity/ChannelEntity.java
@@ -1,5 +1,6 @@
 package com.sprint.mission.discodeit.entity;
 
+import com.sprint.mission.discodeit.domain.ChannelType;
 import jakarta.persistence.*;
 
 @Entity
