+++ src/main/java/com/sprint/mission/discodeit/entity/UserEntity.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.entity;
 
-import com.sprint.mission.discodeit.entity.exception.ErrorType;
-import com.sprint.mission.discodeit.entity.exception.ValidationException;
+import com.sprint.mission.discodeit.domain.exception.ErrorType;
+import com.sprint.mission.discodeit.domain.exception.ValidationException;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -30,22 +30,6 @@ public class UserEntity extends BaseUpdatableEntity {
     private UUID profileId;
 
 
-    private void validateEmail(String email) {
-        if (!email.matches("^[A-Za-z0-9+_.-]+@(.+)$")) {
-            throw new ValidationException(ErrorType.INVALID_EMAIL);
-        }
-    }
 
-    private void validatePassword(String password) {
-        if (password == null || password.length() < 4) {
-            throw new ValidationException(ErrorType.INVALID_PASSWORD);
-        }
-    }
-
-    private void validateUsername(String username) {
-        if (username == null || username.isBlank()) {
-            throw new ValidationException(ErrorType.INVALID_USERNAME);
-        }
-    }
 
 }
