+++ src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -2,7 +2,7 @@ package com.sprint.mission.discodeit.service;
 
 
 import com.sprint.mission.discodeit.domain.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.domain.ChannelType;
 
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.service.dto.request.PublicChannelUpdateRequest;
