+++ src/main/java/com/sprint/mission/discodeit/service/dto/response/ChannelDto.java
@@ -1,0 +1,23 @@
+package com.sprint.mission.discodeit.service.dto.response;
+
+import com.sprint.mission.discodeit.domain.ChannelType;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@Setter
+@Getter
+public class ChannelDto {
+
+    private UUID id;
+    private String name;
+    private String description;
+    private ChannelType type;
+    private List<UserDto> participants = new ArrayList<>();
+    private Instant lastMessageAt;
+
+}
