--- src/main/java/com/sprint/mission/discodeit/service/dto/response/ChannelType.java
@@ -1,5 +1,0 @@
-package com.sprint.mission.discodeit.service.dto.response;
-
-public enum ChannelType {
-    PUBLIC, PRIVATE
-}
