+++ src/main/java/com/sprint/mission/discodeit/service/dto/response/MessageDto.java
@@ -1,0 +1,22 @@
+package com.sprint.mission.discodeit.service.dto.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@Getter
+@Setter
+public class MessageDto {
+
+    private UUID id;
+    private Instant createdAt;
+    private Instant updatedAt;
+    private UUID channelId;
+    private UserDto author;
+    private final List<BinaryContent> attachments = new ArrayList<>();
+
+}
