--- src/main/java/com/sprint/mission/discodeit/service/dto/response/ReadStatus.java
@@ -1,39 +1,0 @@
-package com.sprint.mission.discodeit.service.dto.response;
-
-import lombok.Getter;
-
-import java.time.Duration;
-import java.time.Instant;
-import java.util.UUID;
-
-
-@Getter
-public class ReadStatus {
-
-    private UUID id;
-    private Instant createdAt;
-    private Instant updatedAt;
-    private UUID userId;
-    private UUID channelId;
-    private Instant lastReadAt;
-
-
-    public ReadStatus(UUID userId, UUID channelId, Instant lastReadAt) {
-
-        this.userId = userId;
-        this.channelId = channelId;
-        this.lastReadAt = lastReadAt;
-
-    }
-
-    public void read(){
-        this.lastReadAt=Instant.now();
-    }
-
-    public Long timeSinceLastRead() {
-        if (lastReadAt == null) {
-            return null;
-        }
-        return Duration.between(lastReadAt, Instant.now()).toMinutes();
-    }
-}
